# [monodroid] set MONO_AOT_MODE_NONE when AOT is off